doodle-vision-neural-net/requirements.txt
tensorflow>=2.10.0
keras>=2.10.0
numpy>=1.21.0
matplotlib>=3.5.0
pillow>=9.0.0
scikit-learn>=1.1.0
quickdraw>=0.2.0
tk>=0.1.0 